update detected words ui
